Updated to unity compatability.
Note: currently working on unity-side solution for tests bugging out system.
